moved assessment summary to application details folder
