+++ sources/packages/web/src/router/AESTRoutes.ts
@@ -40,7 +40,7 @@ import AESTApplicationSideBar from "@/components/layouts/aest/AESTApplicationSid
 import SupportingUser from "@/views/aest/SupportingUser.vue";
 import PendingDesignations from "@/views/aest/institution/PendingDesignation.vue";
 import DesignationAESTView from "@/views/aest/institution/DesignationAESTView.vue";
-import AssessmentsSummary from "@/views/aest/AssessmentsSummary.vue";
+import AssessmentsSummary from "@/views/aest/student/applicationDetails/AssessmentsSummary.vue";
 import StudentAppealRequestsApproval from "@/views/aest/student/applicationDetails/StudentAppealRequestsApproval.vue";
 import NoticeOfAssessment from "@/views/aest/NoticeOfAssessment.vue";
 import ApplicationExceptionsApproval from "@/views/aest/student/applicationDetails/ApplicationExceptionsApproval.vue";
