+++ sources/packages/web/src/views/aest/student/applicationDetails/AssessmentsSummary.vue
